Dashboard/requirements.txt
dash
pandas
dash_bootstrap_components
lightgbm
plotly.express
gunicorn
scikit-learn
requests